requirements.txt
python-telegram-bot==20.8
supabase==2.5.1
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.1